+++ Source/Scene/Instanced3DModel3DTileContent.js
@@ -26,6 +26,7 @@ define([
         './Cesium3DTileBatchTable',
         './Cesium3DTileFeature',
         './Cesium3DTileFeatureTable',
+        './ClippingPlaneCollection',
         './ModelInstanceCollection'
     ], function(
         AttributeCompression,
@@ -55,6 +56,7 @@ define([
         Cesium3DTileBatchTable,
         Cesium3DTileFeature,
         Cesium3DTileFeatureTable,
+        ClippingPlaneCollection,
         ModelInstanceCollection) {
     'use strict';
 
@@ -513,6 +515,18 @@ define([
         this._modelInstanceCollection.shadows = this._tileset.shadows;
         this._modelInstanceCollection.debugWireframe = this._tileset.debugWireframe;
         this._modelInstanceCollection.update(frameState);
+
+        // Update clipping planes
+        var tilesetClippingPlanes = this._tileset.clippingPlanes;
+        if (defined(tilesetClippingPlanes)) {
+            var model = this._modelInstanceCollection._model;
+            if (!defined(model.clippingPlanes)) {
+                model.clippingPlanes = new ClippingPlaneCollection();
+            }
+
+            tilesetClippingPlanes.clone(model.clippingPlanes);
+            model.clippingPlanes.enabled = tilesetClippingPlanes.enabled && this._tile._isClipped;
+        }
 
         // If any commands were pushed, add derived commands
         var commandEnd = frameState.commandList.length;