+++ Source/Shaders/Builtin/Constants/maxClippingPlanes.glsl
@@ -1,0 +1,8 @@
+/**
+ * The maximum amount of planes that can be clipped by czm_clipPlanes.
+ *
+ * @name czm_maxClippingPlanes
+ * @glslConstant
+ * @see czm_clipPlanes
+ */
+const int czm_maxClippingPlanes = 6;