+++ Specs/Scene/Instanced3DModel3DTileContentSpec.js
@@ -3,7 +3,10 @@ defineSuite([
         'Core/Color',
         'Core/HeadingPitchRange',
         'Core/HeadingPitchRoll',
+        'Core/Plane',
         'Core/Transforms',
+        'Scene/ClippingPlaneCollection',
+        'Scene/TileBoundingSphere',
         'Specs/Cesium3DTilesTester',
         'Specs/createScene'
     ], function(
@@ -11,7 +14,10 @@ defineSuite([
         Color,
         HeadingPitchRange,
         HeadingPitchRoll,
+        Plane,
         Transforms,
+        ClippingPlaneCollection,
+        TileBoundingSphere,
         Cesium3DTilesTester,
         createScene) {
     'use strict';
@@ -296,6 +302,32 @@ defineSuite([
             expect(content.geometryByteLength).toEqual(geometryByteLength);
             expect(content.texturesByteLength).toEqual(texturesByteLength);
             expect(content.batchTableByteLength).toEqual(batchTexturesByteLength + pickTexturesByteLength);
+        });
+    });
+
+    it('Updates model\'s clipping planes', function() {
+        return Cesium3DTilesTester.loadTileset(scene, withBatchTableUrl).then(function(tileset) {
+            var tile = tileset._root;
+            var content = tile.content;
+            var model = content._modelInstanceCollection._model;
+
+            expect(model.clippingPlanes).toBeUndefined();
+
+            tileset.clippingPlanes = new ClippingPlaneCollection({
+                planes : [
+                    new Plane(Cartesian3.UNIT_X, 0.0)
+                ]
+            });
+            content.update(tileset, scene.frameState);
+
+            expect(model.clippingPlanes).toBeDefined();
+            expect(model.clippingPlanes.planes.length).toBe(1);
+            expect(model.clippingPlanes.enabled).toBe(true);
+
+            tile._isClipped = false;
+            content.update(tileset, scene.frameState);
+
+            expect(model.clippingPlanes.enabled).toBe(false);
         });
     });